:recycle: :: Community Filter Popup에 있는 BlocProvider 위치 변경
